stock-report-writer/requirements.txt
google-generativeai>=0.8.0
google-cloud-texttospeech>=2.16.0
moviepy>=1.0.3
Pillow>=10.0.0
python-dotenv>=1.0.0
